+++ apps/docs-utils/src/configs/menu.ts
@@ -1,18 +1,36 @@
 import type { LayoutMenu } from '@pikas/docs-ui';
-import { docsUtilsRoutes } from '@pikas/docs-routes';
-const { getLink } = docsUtilsRoutes;
+import { docsUtilsRoutes, docsUIRoutes } from '@pikas/docs-routes';
+const { getLink: getLinkUtils } = docsUtilsRoutes;
+const { getLink: getLinkUI } = docsUIRoutes;
 
 export const menu: LayoutMenu = [
   {
     label: 'Docs',
     items: [
       {
-        label: 'Screen',
-        href: getLink('screen'),
+        label: 'File',
+        href: getLinkUtils('file'),
+      },
+      {
+        label: 'Keyboard',
+        href: getLinkUtils('keyboard'),
       },
       {
         label: 'Router',
-        href: getLink('router'),
+        href: getLinkUtils('router'),
+      },
+      {
+        label: 'Screen',
+        href: getLinkUtils('screen'),
+      },
+    ],
+  },
+  {
+    label: 'Pikas',
+    items: [
+      {
+        label: 'Pikas-UI',
+        href: getLinkUI('introduction', { withOrigin: true }),
       },
     ],
   },