+++ apps/docs-utils/src/examples/screen/useMediaScreen/UseMediaScreenExample.tsx
@@ -1,9 +1,19 @@
-import { ExampleContainer } from '@pikas/docs-ui';
+import { ExampleContainer, styled } from '@pikas/docs-ui';
 import { useMediaScreen } from '@pikas-utils/screen';
 import { FC } from 'react';
+
+const Content = styled('div', {
+  color: '$BLACK',
+});
 
 export const UseMediaScreenExample: FC = () => {
   const media = useMediaScreen();
 
-  return <ExampleContainer>{media}</ExampleContainer>;
+  return (
+    <ExampleContainer>
+      <Content>
+        The current media screen is: <b>{media}</b>
+      </Content>
+    </ExampleContainer>
+  );
 };