+++ packages/utils/file/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pikas-utils/file
 
+## 1.0.3
+
+### Patch Changes
+
+- Update package.json
+
 ## 1.0.2
 
 ### Patch Changes