+++ packages/utils/keyboard/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pikas-utils/keyboard
 
+## 1.0.3
+
+### Patch Changes
+
+- Update package.json
+
 ## 1.0.2
 
 ### Patch Changes